requirements.txt
"flet[all]"
pytubefix
ffmpeg-python